# app/Http/Controllers/Api/AuthApiController.php
<?php

    namespace App\Http\Controllers\Api;

    use App\Models\User;
    use Illuminate\Http\Request;
    use Illuminate\Support\Facades\Hash;
    use OpenApi\Annotations as OA;

    class AuthApiController extends ApiController
    {
        /**
         * @OA\Post(
         *     path="/auth/register",
         *     operationId="registerUser",
         *     summary="Register a new user",
         *     tags={"Auth"},
         *     @OA\RequestBody(
         *         required=true,
         *         @OA\JsonContent(
         *             @OA\Property(property="name", type="string", example="John Doe"),
         *             @OA\Property(property="email", type="string", example="john.doe@example.com"),
         *             @OA\Property(property="password", type="string", example="password"),
         *             @OA\Property(property="password_confirmation", type="string", example="password")
         *         )
         *     ),
         *     @OA\Response(
         *         response=201,
         *         description="User successfully registered",
         *         @OA\JsonContent(
         *             @OA\Property(property="message", type="string", example="User successfully registered"),
         *             @OA\Property(property="user", ref="#/components/schemas/User")
         *         )
         *     )
         * )
         */
        public function register(Request $request)
        {
            $validated = $request->validate([
                'name' => 'required|string|max:255',
                'email' => 'required|string|email|max:255|unique:users',
                'password' => 'required|string|min:6|confirmed',
            ]);

            $user = User::create([
                'name' => $validated['name'],
                'email' => $validated['email'],
                'password' => Hash::make($validated['password']),
            ]);

            return response()->json([
                'message' => 'User successfully registered',
                'user' => $user
            ], 201);
        }

        /**
         * @OA\Post(
         *     path="/auth/login",
         *     operationId="loginUser",
         *     summary="Login a user and return a JWT token",
         *     tags={"Auth"},
         *     @OA\RequestBody(
         *         required=true,
         *         @OA\JsonContent(
         *             @OA\Property(property="email", type="string", example="john.doe@example.com"),
         *             @OA\Property(property="password", type="string", example="password")
         *         )
         *     ),
         *     @OA\Response(
         *         response=200,
         *         description="User successfully logged in",
         *         @OA\JsonContent(
         *             @OA\Property(property="access_token", type="string"),
         *             @OA\Property(property="token_type", type="string", example="bearer"),
         *             @OA\Property(property="expires_in", type="integer")
         *         )
         *     ),
         *     @OA\Response(
         *         response=401,
         *         description="Unauthorized",
         *         @OA\JsonContent(
         *             @OA\Property(property="error", type="string", example="Unauthorized")
         *         )
         *     )
         * )
         */
        public function login(Request $request)
        {
            $credentials = $request->only('email', 'password');

            if (! $token = auth()->attempt($credentials)) {
                return response()->json(['error' => 'Unauthorized'], 401);
            }

            return $this->respondWithToken($token);
        }

        /**
         * @OA\Post(
         *     path="/auth/logout",
         *     operationId="logoutUser",
         *     summary="Logout a user",
         *     tags={"Auth"},
         *     @OA\Response(
         *         response=200,
         *         description="Successfully logged out",
         *         @OA\JsonContent(
         *             @OA\Property(property="message", type="string", example="Successfully logged out")
         *         )
         *     )
         * )
         */
        public function logout()
        {
            auth()->logout();
            return response()->json(['message' => 'Successfully logged out']);
        }

        /**
         * @OA\Post(
         *     path="/auth/refresh",
         *     operationId="refreshToken",
         *     summary="Refresh a JWT token",
         *     tags={"Auth"},
         *     @OA\Response(
         *         response=200,
         *         description="Token refreshed successfully",
         *         @OA\JsonContent(
         *             @OA\Property(property="access_token", type="string"),
         *             @OA\Property(property="token_type", type="string", example="bearer"),
         *             @OA\Property(property="expires_in", type="integer")
         *         )
         *     )
         * )
         */
        public function refresh()
        {
            return $this->respondWithToken(auth()->refresh());
        }

        /**
         * @OA\Get(
         *     path="/auth/me",
         *     operationId="getAuthenticatedUser",
         *     summary="Get the authenticated user's information",
         *     tags={"Auth"},
         *     @OA\Response(
         *         response=200,
         *         description="Authenticated user data",
         *         @OA\JsonContent(ref="#/components/schemas/User")
         *     )
         * )
         */
        public function me()
        {
            return response()->json(auth()->user());
        }

        /**
         * Helper function to return token response
         */
        protected function respondWithToken($token)
        {
            return response()->json([
                'access_token' => $token,
                'token_type' => 'bearer',
                'expires_in' => auth()->factory()->getTTL() * 60
            ]);
        }
    }